// src/types.ts
export interface Tag {
  slug: string;
  label: string;
  url: string;
  count: number;
}

export interface Category {
  slug: string;
  label: string;
  url: string;
  count: number;
}

export interface Author {
  name: string;
  avatar?: string | ImageMetadata | Promise<{ default: ImageMetadata }>;
  description?: string;
  twitter?: string;
  github?: string;
  linkedin?: string;
}

export interface Card {
  tokenId: string;
  name: string;
  description: string;
  image: string;
  attributes: Record<string, string>;
  external_url: string;
  artist: string;
  collection: string;
  conservation_status: string;
}

export interface Post {
  slug: string;
  url: string;
  title: string;
  date: Date;
  updateDate: Date;
  draft: boolean;
  category: Category;
  tags: Tag[];
  author: Author;
  excerpt?: string;
  image?: string | ImageMetadata | Promise<{ default: ImageMetadata }>;
  permalink: string;
  readingTime?: number;
  wordCount?: number;
  cardVariant?: "blur" | "material" | "full" | "plain";
  raw: string;
  headings: MarkdownHeading[];
  Content: MarkdownInstance<Record<string, unknown>>["Content"];
}